Rename workflow::handlers the workflow::script module
The script definition has been moved to the tedge_api::script module
And the former workflow::script module contains now only workflow
handlers.
