+++ crates/core/tedge_api/src/workflow/mod.rs
@@ -1,7 +1,7 @@
 pub mod error;
+pub mod handlers;
 pub(crate) mod log;
 mod on_disk;
-pub mod script;
 pub mod state;
 pub mod supervisor;
 mod toml_config;
@@ -12,9 +12,9 @@ use crate::mqtt_topics::OperationType;
 use crate::script::ShellScript;
 use ::log::info;
 pub use error::*;
+pub use handlers::*;
 use mqtt_channel::MqttMessage;
 use mqtt_channel::QoS;
-pub use script::*;
 use serde::Deserialize;
 use serde_json::json;
 pub use state::*;
